claude: refactor(kaspa): rename kas_bridge to bridge
Rename kas_bridge module to simply 'bridge' - clearer semantic naming
that describes the module's purpose (bridge operation data types:
DepositFXG, WithdrawFXG, ConfirmationFXG, message parsing).
