+++ rust/main/chains/dymension-kaspa/src/bridge/withdraw.rs
@@ -1,4 +1,4 @@
-use crate::kas_bridge::payload::MessageID;
+use crate::bridge::payload::MessageID;
 use bytes::Bytes;
 use eyre::Error as EyreError;
 use hex::ToHex;
