+++ rust/main/chains/dymension-kaspa/src/kas_validator/withdraw_test.rs
@@ -4,9 +4,9 @@
 //     use crate::kas_validator::error::ValidationError;
 //     use crate::withdraw::validate_pskts;
 //     use dym_kas_core::escrow::{Escrow, EscrowPublic};
-//     use crate::kas_bridge::payload::{MessageID, MessageIDs};
-//     use crate::kas_bridge::util::get_recipient_script_pubkey_address;
-//     use crate::kas_bridge::withdraw::WithdrawFXG;
+//     use crate::bridge::payload::{MessageID, MessageIDs};
+//     use crate::bridge::util::get_recipient_script_pubkey_address;
+//     use crate::bridge::withdraw::WithdrawFXG;
 //     use eyre::{eyre, Result}; // Use eyre's Result for convenient error handling
 //     use hyperlane_core::{Encode, HyperlaneMessage, H256, U256};
 //     use hyperlane_warp_route::TokenMessage;
