+++ rust/main/chains/dymension-kaspa/src/providers/provider.rs
@@ -1,7 +1,7 @@
 use super::confirmation_queue::PendingConfirmation;
 use super::validators::ValidatorsClient;
 use super::RestProvider;
-use crate::kas_bridge::confirmation::ConfirmationFXG;
+use crate::bridge::confirmation::ConfirmationFXG;
 use crate::kas_relayer::withdraw::hub_to_kaspa::combine_bundles_with_fee;
 use crate::kas_relayer::withdraw::messages::on_new_withdrawals;
 use crate::kas_relayer::KaspaBridgeMetrics;
@@ -35,7 +35,7 @@ use tracing::{error, info};
 use url::Url;
 
 struct ProcessedWithdrawals {
-    fxg: std::sync::Arc<crate::kas_bridge::withdraw::WithdrawFXG>,
+    fxg: std::sync::Arc<crate::bridge::withdraw::WithdrawFXG>,
     tx_ids: Vec<RpcTransactionId>,
 }
